+++ src/components/Multiselect/__tests__/Multiselect.spec.js
@@ -69,7 +69,7 @@ describe('Multiselect', () => {
       const { findByTestId } = component;
 
       const inputContainer = await findByTestId('input-container');
-      expect(inputContainer).toHaveClass('focus-within:ring-1 focus-within:ring-primary-500');
+      expect(inputContainer).toHaveClass('focus-within:border-blue-500');
     });
 
     it('renders a dropdown of options', async () => {